Integration tests (#17)
Added support for a test server which sets up a server used for testing
purposes. It will test the full stack from rest endpoints to the
postgres database.

* Remove pq from vendor
Removed pq from vendor since it creates troubles testing sub packages.
It results in two versions of pq is imported and they conflict.

* Fix blob repository tests
* Example integration test
The test doesn't really test anything meaningful and should probably be
rewritten. But it is now possible to set up a test server

* Example test for postgres repository
Example test for the repositories to test whether data is being inserted
and fetched correctly from the underlying database